big big feat: Add Settings Page
>Removed Your Themes Modal
>Removed "Installed Themes" Tab
>The functionality of both of those have been moved into the "Themes" section of the new settings menu

>Moved plugin settings out of the store settings page into the settings tab in the new settings page

>Moved credits to the credits tab of the settings page

>Theme errors are displayed in the Themes tab

>many more changes but i am very tire
