+++ src/components/Modals/index.ts
@@ -1,4 +1,2 @@
-export * from "./AllThemesModal";
 export * from "./CreatePresetModal";
 export * from "./ThemeSettingsModal";
-export * from "./AllThemesSingleEntry";
