+++ src/components/index.ts
@@ -3,6 +3,6 @@ export * from "./ThemePatch";
 export * from "./PatchComponent";
 export * from "./TitleView";
 export * from "./ThemeManager";
-export * from "./AllThemes";
 export * from "./OptionalDepsModal";
 export * from "./QAMTab";
+export * from "./Modals";
