+++ src/pages/settings/Credits.tsx
@@ -1,0 +1,40 @@
+export function Credits() {
+  return (
+    <div>
+      <div style={{ display: "flex", flexDirection: "column", fontSize: "0.9em" }}>
+        <div>
+          <h2 style={{ fontWeight: "bold", fontSize: "1.5em", marginBottom: "0px" }}>Developers</h2>
+          <ul style={{ marginTop: "0px", marginBottom: "0px" }}>
+            <li>
+              <span>SuchMeme - github.com/suchmememanyskill</span>
+            </li>
+            <li>
+              <span>Beebles - github.com/beebls</span>
+            </li>
+            <li>
+              <span>EMERALD - github.com/EMERALD0874</span>
+            </li>
+          </ul>
+        </div>
+        <div>
+          <h2 style={{ fontWeight: "bold", fontSize: "1.5em", marginBottom: "0px" }}>Support</h2>
+          <span>
+            See the DeckThemes Discord server for support.
+            <br />
+            discord.gg/HsU72Kfnpf
+          </span>
+        </div>
+        <div>
+          <h2 style={{ fontWeight: "bold", fontSize: "1.5em", marginBottom: "0px" }}>
+            Create and Submit Your Own Theme
+          </h2>
+          <span>
+            Instructions for theme creation/submission are available DeckThemes' docs website.
+            <br />
+            docs.deckthemes.com
+          </span>
+        </div>
+      </div>
+    </div>
+  );
+}
