+++ src/pages/theme-manager/SubmissionBrowserPage.tsx
@@ -1,46 +1,33 @@
 import { Focusable } from "decky-frontend-lib";
-import { useLayoutEffect, useState, FC, useEffect, useRef } from "react";
-import * as python from "../python";
-import { getThemes } from "../api";
-import { logInWithShortToken } from "../api";
+import { useCssLoaderState } from "../../state";
+import * as python from "../../python";
+import { BrowserSearchFields, LoadMoreButton, VariableSizeCard } from "../../components";
+import { useEffect, useRef, useState } from "react";
 import { isEqual } from "lodash";
+import { getThemes } from "../../api";
 
-// Interfaces for the JSON objects the lists work with
-import { useCssLoaderState } from "../state";
-import { BrowserSearchFields, VariableSizeCard, LoadMoreButton } from "../components";
-
-export const ThemeBrowserPage: FC = () => {
+export function SubmissionsPage() {
   const {
-    browseThemeList: themeArr,
-    themeSearchOpts: searchOpts,
-    apiShortToken,
     apiFullToken,
-    serverFilters,
-    browserCardSize = 3,
-    prevSearchOpts,
+    submissionSearchOpts: searchOpts,
+    submissionServerFilters: serverFilters,
+    submissionThemeList: themeArr,
+    browserCardSize,
+    prevSubSearchOpts: prevSearchOpts,
+    apiMeData,
     backendVersion,
   } = useCssLoaderState();
 
   function reloadThemes() {
-    getThemes(searchOpts, "/themes", "browseThemeList", setSnapIndex);
+    getThemes(searchOpts, "/themes/awaiting_approval", "submissionThemeList", setSnapIndex, true);
     python.reloadBackend();
   }
 
   useEffect(() => {
     if (!isEqual(prevSearchOpts, searchOpts) || themeArr.total === 0) {
-      getThemes(searchOpts, "/themes", "browseThemeList", setSnapIndex);
+      getThemes(searchOpts, "/themes/awaiting_approval", "submissionThemeList", setSnapIndex, true);
     }
-  }, [searchOpts, prevSearchOpts]);
-
-  // Runs upon opening the page every time
-  useLayoutEffect(() => {
-    python.getBackendVersion();
-    if (apiShortToken && !apiFullToken) {
-      logInWithShortToken();
-    }
-    // Installed themes aren't used on this page, but they are used on other pages, so fetching them here means that as you navigate to the others they will be already loaded
-    python.getInstalledThemes();
-  }, []);
+  }, [searchOpts, prevSearchOpts, apiMeData]);
 
   const endOfPageRef = useRef<HTMLElement>();
   const [indexToSnapTo, setSnapIndex] = useState<number>(-1);
@@ -50,20 +37,36 @@ export const ThemeBrowserPage: FC = () => {
     }
   }, [indexToSnapTo]);
 
+  if (!apiFullToken) {
+    return (
+      <>
+        <div
+          style={{
+            display: "flex",
+            flexDirection: "column",
+            justifyContent: "center",
+            alignItems: "center",
+            height: "80%",
+          }}
+        >
+          <span style={{ fontSize: "1.5em", fontWeight: "bold" }}>You Are Not Logged In!</span>
+          <span>Link your deck to your deckthemes.com account to sync Starred Themes</span>
+        </div>
+      </>
+    );
+  }
   return (
     <>
       <BrowserSearchFields
         searchOpts={searchOpts}
-        searchOptsVarName="themeSearchOpts"
-        prevSearchOptsVarName="prevSearchOpts"
+        searchOptsVarName="submissionSearchOpts"
+        prevSearchOptsVarName="prevSubSearchOpts"
         unformattedFilters={serverFilters}
-        unformattedFiltersVarName="serverFilters"
-        getTargetsPath="/themes/filters?type=CSS"
-        onReload={() => {
-          reloadThemes();
-        }}
+        unformattedFiltersVarName="submissionServerFilters"
+        getTargetsPath="/themes/awaiting_approval/filters?type=CSS"
+        requiresAuth
+        onReload={reloadThemes}
       />
-      {/* I wrap everything in a Focusable, because that ensures that the dpad/stick navigation works correctly */}
       <Focusable
         style={{
           display: "flex",
@@ -82,7 +85,7 @@ export const ThemeBrowserPage: FC = () => {
               cols={browserCardSize}
               showTarget={true}
               searchOpts={searchOpts}
-              prevSearchOptsVarName="prevSearchOpts"
+              prevSearchOptsVarName="setPrevSubSearchOpts"
             />
           ))}
       </Focusable>
@@ -98,14 +101,14 @@ export const ThemeBrowserPage: FC = () => {
         <div style={{ maxWidth: "50%" }}>
           <LoadMoreButton
             themeArr={themeArr}
-            themeArrVarName="browseThemeList"
+            themeArrVarName="submissionThemeList"
             origSearchOpts={searchOpts}
             paramStrFilterPrepend="CSS."
-            fetchPath="/themes"
+            fetchPath="/themes/awaiting_approval"
             setSnapIndex={setSnapIndex}
           />
         </div>
       </div>
     </>
   );
-};
+}
