+++ src/pages/theme-manager/ThemeBrowserPage.tsx
@@ -1,33 +1,46 @@
 import { Focusable } from "decky-frontend-lib";
-import { useCssLoaderState } from "../state";
-import * as python from "../python";
-import { BrowserSearchFields, LoadMoreButton, VariableSizeCard } from "../components";
-import { useEffect, useRef, useState } from "react";
+import { useLayoutEffect, useState, FC, useEffect, useRef } from "react";
+import * as python from "../../python";
+import { getThemes } from "../../api";
+import { logInWithShortToken } from "../../api";
 import { isEqual } from "lodash";
-import { getThemes } from "../api";
 
-export function StarredThemesPage() {
+// Interfaces for the JSON objects the lists work with
+import { useCssLoaderState } from "../../state";
+import { BrowserSearchFields, VariableSizeCard, LoadMoreButton } from "../../components";
+
+export const ThemeBrowserPage: FC = () => {
   const {
+    browseThemeList: themeArr,
+    themeSearchOpts: searchOpts,
+    apiShortToken,
     apiFullToken,
-    apiMeData,
-    starredSearchOpts: searchOpts,
-    starredServerFilters: serverFilters,
-    starredThemeList: themeArr,
-    browserCardSize,
-    prevStarSearchOpts: prevSearchOpts,
+    serverFilters,
+    browserCardSize = 3,
+    prevSearchOpts,
     backendVersion,
   } = useCssLoaderState();
 
   function reloadThemes() {
-    getThemes(searchOpts, "/users/me/stars", "starredThemeList", setSnapIndex, true);
+    getThemes(searchOpts, "/themes", "browseThemeList", setSnapIndex);
     python.reloadBackend();
   }
 
   useEffect(() => {
     if (!isEqual(prevSearchOpts, searchOpts) || themeArr.total === 0) {
-      getThemes(searchOpts, "/users/me/stars", "starredThemeList", setSnapIndex, true);
+      getThemes(searchOpts, "/themes", "browseThemeList", setSnapIndex);
     }
-  }, [searchOpts, prevSearchOpts, apiMeData]);
+  }, [searchOpts, prevSearchOpts]);
+
+  // Runs upon opening the page every time
+  useLayoutEffect(() => {
+    python.getBackendVersion();
+    if (apiShortToken && !apiFullToken) {
+      logInWithShortToken();
+    }
+    // Installed themes aren't used on this page, but they are used on other pages, so fetching them here means that as you navigate to the others they will be already loaded
+    python.getInstalledThemes();
+  }, []);
 
   const endOfPageRef = useRef<HTMLElement>();
   const [indexToSnapTo, setSnapIndex] = useState<number>(-1);
@@ -37,36 +50,20 @@ export function StarredThemesPage() {
     }
   }, [indexToSnapTo]);
 
-  if (!apiFullToken) {
-    return (
-      <>
-        <div
-          style={{
-            display: "flex",
-            flexDirection: "column",
-            justifyContent: "center",
-            alignItems: "center",
-            height: "80%",
-          }}
-        >
-          <span style={{ fontSize: "1.5em", fontWeight: "bold" }}>You Are Not Logged In!</span>
-          <span>Link your deck to your deckthemes.com account to sync Starred Themes</span>
-        </div>
-      </>
-    );
-  }
   return (
     <>
       <BrowserSearchFields
         searchOpts={searchOpts}
-        searchOptsVarName="starredSearchOpts"
-        prevSearchOptsVarName="prevStarSearchOpts"
+        searchOptsVarName="themeSearchOpts"
+        prevSearchOptsVarName="prevSearchOpts"
         unformattedFilters={serverFilters}
-        unformattedFiltersVarName="starredServerFilters"
-        getTargetsPath="/users/me/stars/filters?type=CSS"
-        requiresAuth
-        onReload={reloadThemes}
+        unformattedFiltersVarName="serverFilters"
+        getTargetsPath="/themes/filters?type=CSS"
+        onReload={() => {
+          reloadThemes();
+        }}
       />
+      {/* I wrap everything in a Focusable, because that ensures that the dpad/stick navigation works correctly */}
       <Focusable
         style={{
           display: "flex",
@@ -85,7 +82,7 @@ export function StarredThemesPage() {
               cols={browserCardSize}
               showTarget={true}
               searchOpts={searchOpts}
-              prevSearchOptsVarName="prevStarSearchOpts"
+              prevSearchOptsVarName="prevSearchOpts"
             />
           ))}
       </Focusable>
@@ -101,14 +98,14 @@ export function StarredThemesPage() {
         <div style={{ maxWidth: "50%" }}>
           <LoadMoreButton
             themeArr={themeArr}
-            themeArrVarName="starredThemeList"
+            themeArrVarName="browseThemeList"
             origSearchOpts={searchOpts}
             paramStrFilterPrepend="CSS."
-            fetchPath="/users/me/stars"
+            fetchPath="/themes"
             setSnapIndex={setSnapIndex}
           />
         </div>
       </div>
     </>
   );
-}
+};
