+++ src/pages/theme-manager/index.ts
@@ -1,7 +1,6 @@
 export * from "./ThemeBrowserPage";
-export * from "./UninstallThemePage";
 export * from "./ExpandedView";
-export * from "./SettingsPage";
+export * from "./LogInPage";
 export * from "./StarredThemesPage";
 export * from "./SubmissionBrowserPage";
 export * from "./ThemeManagerRouter";
